feat: policies on all modules

+++ app/Http/Controllers/Admin/VetAppointmentScheduleController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\Admin\UpsertVetAppointmentScheduleRequest;
 use App\Models\Admin;
 use App\Models\VetAppointmentSchedule;
 use App\Models\VetService;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class VetAppointmentScheduleController extends Controller
@@ -16,6 +17,7 @@ class VetAppointmentScheduleController extends Controller
 	 */
 	public function index()
 	{
+		Gate::authorize('viewAny', VetAppointmentSchedule::class);
 		return Inertia::render('admin/VetAppointmentSchedules', [
 			'pagination' => Inertia::always(Inertia::merge(VetAppointmentSchedule::with(['service', 'doctor'])->orderBy(request('sortField', 'created_at'), request('sortDirection', 'desc'))->paginate(request('perPage', 5), "*", null, request('page', 1)))),
 			'vetServices' => Inertia::lazy(fn() => VetService::all()),
@@ -30,6 +32,7 @@ class VetAppointmentScheduleController extends Controller
 	 */
 	public function store(UpsertVetAppointmentScheduleRequest $request)
 	{
+		Gate::authorize('store', VetAppointmentSchedule::class);
 		VetAppointmentSchedule::create($request->all());
 
 		return redirect('/admin/vet-appointment-schedules');
@@ -40,6 +43,7 @@ class VetAppointmentScheduleController extends Controller
 	 */
 	public function update(UpsertVetAppointmentScheduleRequest $request, string $id)
 	{
+		Gate::authorize('update', VetAppointmentSchedule::class);
 		$data = VetAppointmentSchedule::findOrFail($id);
 		$data->update($request->all());
 
@@ -51,6 +55,7 @@ class VetAppointmentScheduleController extends Controller
 	 */
 	public function destroy(string $id)
 	{
+		Gate::authorize('delete', VetAppointmentSchedule::class);
 		$row = VetAppointmentSchedule::findOrFail($id);
 		$row->delete();
