+++ app/Http/Controllers/Admin/VetServiceController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\UpsertVetServiceRequest;
 use App\Models\Upload;
+use Illuminate\Support\Facades\Gate;
 use App\Models\VetService;
 use Illuminate\Support\Facades\DB;
 use Inertia\Inertia;
@@ -17,6 +18,7 @@ class VetServiceController extends Controller
 	 */
 	public function index()
 	{
+		Gate::authorize('viewAny', VetService::class);
 		return Inertia::render('admin/VetServices', [
 			'pagination' => Inertia::always(Inertia::merge(VetService::with('uploads', 'types')->orderBy(request('sortField', 'created_at'), request('sortDirection', 'desc'))->paginate(request('perPage', 5), "*", null, request('page', 1)))),
 		]);
@@ -24,6 +26,7 @@ class VetServiceController extends Controller
 
 	public function store(UpsertVetServiceRequest $request)
 	{
+		Gate::authorize('store', VetService::class);
 		DB::transaction(function () use ($request) {
 			$vetService = VetService::create($request->all());
 
@@ -48,6 +51,7 @@ class VetServiceController extends Controller
 
 	public function update(UpsertVetServiceRequest $request, string $id)
 	{
+		Gate::authorize('update', VetService::class);
 		DB::transaction(function () use ($id, $request) {
 			$data = VetService::findOrFail($id);
 			$data->update($request->all());
@@ -73,6 +77,7 @@ class VetServiceController extends Controller
 
 	public function destroy(string $id)
 	{
+		Gate::authorize('delete', VetService::class);
 		$row = VetService::findOrFail($id);
 		$row->delete();
 
@@ -81,6 +86,7 @@ class VetServiceController extends Controller
 
 	public function deleteImage(string $id)
 	{
+		Gate::authorize('delete', VetService::class);
 		DB::transaction(function () use ($id, ) {
 			$file = Upload::findOrFail($id);
 			Storage::disk('public')->delete('vet_services/' . $file->file_name);
